// src/shared/api/orval/user-service/user-service.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * User API
 * OpenAPI spec version: 1.0.0
 */
export interface MessageData {
  message: string;
}

export interface BaseResponse {
  status: string;
  data: MessageData;
}

export interface BlacklistCreateRequest {
  user_id?: string | null;
  login?: string | null;
}

export interface BlacklistInfo {
  user_id: string;
  login?: string | null;
  full_name?: string | null;
  /** Дата добавления в Blacklist */
  created_at: string;
}

export interface BlacklistCreateResponse {
  status: string;
  data: BlacklistInfo;
}

export interface BlacklistDeleteRequest {
  user_id: string;
}

export interface BlacklistInfoData {
  items: BlacklistInfo[];
  has_more: boolean;
}

export interface BlacklistInfoResponse {
  status: string;
  data: BlacklistInfoData;
}

export interface ContactCountData {
  count: number;
}

export interface ContactCountResponse {
  status: string;
  data: ContactCountData;
}

export interface ContactCreateRequest {
  user_id?: string | null;
  login?: string | null;
  friend_code?: string | null;
  custom_name?: string | null;
}

export interface ContactInfo {
  user_id: string;
  login?: string | null;
  full_name?: string | null;
  custom_name?: string | null;
  friend_code: boolean;
  /** Дата добавления в контакты */
  created_at: string;
  /** Дата последнего онлайна */
  last_seen_at?: string | null;
}

export interface ContactCreateResponse {
  status: string;
  data: ContactInfo;
}

export interface ContactDeleteRequest {
  user_id: string;
}

export interface ContactInfoData {
  items: ContactInfo[];
  has_more: boolean;
}

export interface ContactInfoResponse {
  status: string;
  data: ContactInfoData;
}

export interface ContactUpdateRequest {
  user_id: string;
  custom_name?: string | null;
}

export interface UserPrivilege {
  id: number;
  code: string;
  name: string;
  description: string;
  is_active: boolean;
  created_at: string;
}

export interface UserRoleData {
  /** ID пользователя */
  user_id: string;
  /** Привилегия пользователя */
  user_privilege?: UserPrivilege | null;
  /** Список кодов прав */
  user_permissions?: string[];
}

export interface UserRoleResponse {
  status: string;
  data: UserRoleData;
}

export type GetContactsContactListGetParams = {
/**
 * Смещение для пагинации
 * @minimum 0
 */
offset?: number;
/**
 * Количество юзеров для загрузки
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type GetBlacklistBlacklistListGetParams = {
/**
 * Смещение для пагинации
 * @minimum 0
 */
offset?: number;
/**
 * Количество юзеров для загрузки
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type GetUserRoleRoleGetParams = {
user_id?: string;
};